Added SUSG01 project data requester and renamed a folder
